tests/wpt/web-platform-tests/html/semantics/forms/the-form-element/form-action-reflection-with-base-url.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>form.action with a base URL</title>
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">
<link rel="help" href="https://html.spec.whatwg.org/multipage/forms.html#dom-fs-action">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<base href="/common/blank.html">

<form id="form1" action="a.html"></form>
<form id="form2" action=""></form>
<form id="form3"></form>

<script>
"use strict";

test(() => {

  assert_equals(document.querySelector("#form1").action, (new URL("a.html", document.baseURI)).href,
    "action should equal the correct absolute URL");

}, "An action URL should be resolved relative to the document's base URL (not the document's URL)");

test(() => {

  assert_equals(document.querySelector("#form2").action, document.URL);

}, "An empty-string action content attribute should cause the IDL attribute to return the document's URL (not the document's base URL)");

test(() => {

  assert_equals(document.querySelector("#form3").action, document.URL);

}, "A missing action content attribute should cause the IDL attribute to return the document's URL (not the document's base URL)");

</script>
