LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-form-element/form-action-reflection-with-base-url-expected.txt


PASS An action URL should be resolved relative to the document's base URL (not the document's URL) 
FAIL An empty-string action content attribute should cause the IDL attribute to return the document's URL (not the document's base URL) assert_equals: expected "http://localhost:8800/html/semantics/forms/the-form-element/form-action-reflection-with-base-url.html" but got "http://localhost:8800/common/blank.html"
FAIL A missing action content attribute should cause the IDL attribute to return the document's URL (not the document's base URL) assert_equals: expected "http://localhost:8800/html/semantics/forms/the-form-element/form-action-reflection-with-base-url.html" but got ""
